requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
openai>=1.3.0
pydantic>=2.0.0
python-dotenv>=1.0.0
httpx>=0.24.0
PyPDF2>=3.0.0
python-docx>=0.8.11
requests>=2.31.0
PyMuPDF>=1.23.0
Pillow>=10.0.0
supabase>=2.0.0
boto3>=1.28.0
